move getRule into StringHelper
